Simplify config source parsing (#25)
Reduce the amount Source suclasses need to know about actions by not
expecting them to combine values that are specified multiple times in an
action specific manner. Instead of the Source subclass combining the
values, it just returns the values as a list to be dealt with later by
the _ConfigSpec subclass for the specific action.