Simplify default handling (#63)
Resolves #62

Previously each action implemented an apply_default() method. That was
done so that the append action would not have to accumulate the default
value into the final value for a config. Since then, I have found out
that Argparse's append action does accumulate the default value into the
final value for a command line parameter and this library has been
updated to match Argparse's behaviour.

For simplicity of implementing new actions (which users might do
themselves) and to make implementing new actions more similar to
implementing Argparse actions, remove the need for an action class to
implement apply_default().